+++ checkdmarc.py
@@ -40,7 +40,7 @@ See the License for the specific language governing permissions and
 limitations under the License."""
 
 
-__version__ = "1.2.0"
+__version__ = "1.2.1"
 
 # Python 2 comparability hack
 if version_info[0] >= 3:
@@ -210,7 +210,7 @@ spf_qualifiers = {
 }
 
 
-def _get_mx_hosts(domain, nameservers=None, timeout=1.0):
+def _get_mx_hosts(domain, nameservers=None, timeout=4.0):
     """
     Queries DNS for a list of Mail Exchange hosts
 
@@ -240,7 +240,7 @@ def _get_mx_hosts(domain, nameservers=None, timeout=1.0):
     return hosts
 
 
-def _get_a_records(domain, nameservers=None, timeout=1.0):
+def _get_a_records(domain, nameservers=None, timeout=4.0):
     """
     Queries DNS for A and AAAA records
 
@@ -278,7 +278,7 @@ def _get_a_records(domain, nameservers=None, timeout=1.0):
     return records
 
 
-def _get_txt_records(domain, nameservers=None, timeout=1.0):
+def _get_txt_records(domain, nameservers=None, timeout=4.0):
     """
     Queries DNS for TXT records
 
@@ -309,7 +309,7 @@ def _get_txt_records(domain, nameservers=None, timeout=1.0):
     return records
 
 
-def _query_dmarc_record(domain, nameservers=None, timeout=1.0):
+def _query_dmarc_record(domain, nameservers=None, timeout=4.0):
     """
     Queries DNS for a DMARC record
 
@@ -340,7 +340,7 @@ def _query_dmarc_record(domain, nameservers=None, timeout=1.0):
     return record
 
 
-def get_mx(domain, nameservers=None, timeout=1.0):
+def get_mx(domain, nameservers=None, timeout=4.0):
     """
     Returns a list of OrderedDicts with keys of ``hostname`` and ``addresses``
     
@@ -373,7 +373,7 @@ def get_mx(domain, nameservers=None, timeout=1.0):
     return OrderedDict([("hosts", hosts), ("warnings", warnings)])
 
 
-def query_dmarc_record(domain, nameservers=None, timeout=1.0):
+def query_dmarc_record(domain, nameservers=None, timeout=4.0):
     """
     Queries DNS for a DMARC record
     Args:
@@ -438,7 +438,7 @@ def parse_dmarc_report_uri(uri):
     return mailto_matches[0]
 
 
-def verify_external_dmarc_destination(source_domain, destination_domain, nameservers=None, timeout=1.0):
+def verify_external_dmarc_destination(source_domain, destination_domain, nameservers=None, timeout=4.0):
     """
       Checks if a report sender is authorized to send a report to the destination domain, per RFC 7489, section 7.1
       
@@ -572,7 +572,7 @@ def parse_dmarc_record(record, organisational_domain, include_tag_descriptions=F
     return OrderedDict([("tags", tags), ("warnings", warnings)])
 
 
-def get_dmarc_record(domain, include_tag_descriptions=False, nameservers=None, timeout=1.0):
+def get_dmarc_record(domain, include_tag_descriptions=False, nameservers=None, timeout=4.0):
     """
     Retrieves a DMARC record for a domain and parses it
 
@@ -595,7 +595,7 @@ def get_dmarc_record(domain, include_tag_descriptions=False, nameservers=None, t
     return OrderedDict([("record", record), ("organisational_domain", organisational_domain), ("tags", tags)])
 
 
-def query_spf_record(domain, nameservers=None, timeout=1.0):
+def query_spf_record(domain, nameservers=None, timeout=4.0):
     """
     Queries DNS for a SPF record
     Args:
@@ -633,7 +633,7 @@ def query_spf_record(domain, nameservers=None, timeout=1.0):
     return spf_record
 
 
-def parse_spf_record(record, domain, seen=None, query_mechanism_count=0, nameservers=None, timeout=1.0):
+def parse_spf_record(record, domain, seen=None, query_mechanism_count=0, nameservers=None, timeout=4.0):
     """
     Parses a SPF record, including resolving ``a``, ``mx``, and ``include`` mechanisms
     
@@ -755,7 +755,7 @@ def parse_spf_record(record, domain, seen=None, query_mechanism_count=0, nameser
     return OrderedDict([("results", results), ("warnings", warnings)])
 
 
-def get_spf_record(domain, query_count=0, nameservers=None, timeout=1.0):
+def get_spf_record(domain, query_count=0, nameservers=None, timeout=4.0):
     """
     Retrieves and parses an SPF record 
     
@@ -776,7 +776,7 @@ def get_spf_record(domain, query_count=0, nameservers=None, timeout=1.0):
 
 
 def check_domains(domains, output_format="json", output_path=None, include_dmarc_tag_descriptions=False,
-                  nameservers=None, timeout=1.0, wait=0.0):
+                  nameservers=None, timeout=4.0, wait=0.0):
     """
     Check the given domains for SPF and DMARC records, parse them, and return them
     
@@ -813,11 +813,11 @@ def check_domains(domains, output_format="json", output_path=None, include_dmarc
         for domain in domains:
             row = dict(domain=domain, mx="", spf_valid=True, dmarc_valid=True)
             mx = get_mx(domain, nameservers=nameservers, timeout=timeout)
-            row["mx"] = " ".join(list(map(lambda r: r["hostname"], mx["hosts"])))
-            row["mx_warnings"] = " ".join(mx["warnings"])
+            row["mx"] = ";".join(list(map(lambda r: r["hostname"], mx["hosts"])))
+            row["mx_warnings"] = ";".join(mx["warnings"])
             try:
                 row["spf_record"] = query_spf_record(domain, nameservers=nameservers, timeout=timeout)
-                row["spf_warnings"] = " ".join(parse_spf_record(row["spf_record"], row["domain"])["warnings"])
+                row["spf_warnings"] = ";".join(parse_spf_record(row["spf_record"], row["domain"])["warnings"])
             except SPFError as error:
                 row["spf_error"] = error
                 row["spf_valid"] = False
@@ -838,7 +838,7 @@ def check_domains(domains, output_format="json", output_path=None, include_dmarc
                     row["dmarc_rua"] = dmarc["tags"]["rua"]["value"]
                 if "ruf" in dmarc:
                     row["dmarc_ruf"] = dmarc["tags"]["ruf"]["value"]
-                row["dmarc_warnings"] = " ".join(dmarc["warnings"])
+                row["dmarc_warnings"] = ";".join(dmarc["warnings"])
             except DMARCError as error:
                 row["dmarc_error"] = error
                 row["dmarc_valid"] = False
@@ -900,8 +900,8 @@ def _main():
     arg_parser.add_argument("-o", "--output", help="output to a file path rather than printing to the screen")
     arg_parser.add_argument("-n", "--nameserver", nargs="+", help="nameservers to query")
     arg_parser.add_argument("-t", "--timeout",
-                            help="number of seconds to wait for an answer from DNS (default 1.0)", type=float,
-                            default=1.0)
+                            help="number of seconds to wait for an answer from DNS (default 4.0)", type=float,
+                            default=4.0)
     arg_parser.add_argument("-v", "--version", action="version", version=__version__)
     arg_parser.add_argument("-w", "--wait", type=float,
                             help="number os seconds to wait between processing domains (default 0.0)",